change qcu-device from qemu-2.4.0 to qemu-2.12.1

+++ qcu-device/roms/SLOF/lib/libc/include/stddef.h
@@ -16,8 +16,7 @@
 
 #define NULL ((void *)0)
 
-
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 
 
 #endif
